Better element rendering and conditionals [skip ci]
Conditional elements are read into javascript literally now, not via
ajax call, so the page should have no delay in responding to clicks and
other ajax-calls because it won't be waiting for the answer to all the
calls it has sent at load time. Also standardizes how elements are
rendered into forms, which is more future proof for any future changes
to how conditionals and rendering work in alternative templates.