Literal nums in linked source queries [skip ci]
Passing a number as the search term when searching on a linked selectbox
used to always be interpretted as a reference to the entry id in the
source form. But with a new global flag, you can have the number treated
as a literal value that should be searched for in the source entries.