Create new field in data table for Foreign keys
Create a new BIGINT(19) field in each form's data table to reflect the
new BIGINT(19) field created in the "formulize" table (aka the elements table).
Possible problems: The field is being updated in the
"relationships_settings_save.php" file by passing in a specific element.
This is creating a new field in each form data table for each respective
element.
ex) For two linking forms A and B, when A links to B, A will
create a new element in B named "elem1", but also in B's data table a new field named "elem1". Now if another form C comes along and links to B and creates a new element in B named "elem2", there will be another field created in B named "elem2". This is possibly a problem and will be changed in the near future most likely.