Fix bug in one to many entries in Foreign key/ Primary key selction
Fix the bug where when Primary key / Foreign key connection, when
clicking on an entry, some of the entry information wouldn't load. This
was due to a problem in the findLinkedEntries() function in the
functions.php file. The problem was the "targetForm['keyself'] == 0"
comparison was comparing the string value of "targetForm['keyself']" to
the integer value 0. When the comparison was made "targetForm['keyself']" was being truncated to a 0 since it was a string value, preventing the new comparisons ("targetForm['keyself'] == 'id'" and "targetForm['keyother'] == 'id'") from being made. The solution was to change the integer value 0 to a string value '0'.