Gather dataset based on 0-index not Assoc [skip ci]
When datasets contain a reference to the same form twice (recursive
one-to-many reference to the main form) then we have to gather the
values from the result object based on a numerical index, not the
associative fetchArray method, because that loses the earlier values and
overwrites them with the later ones. Yuck.