Gathering appropriate frid properly from rendered screen
For use in setting one to one links, etc. One to ones were broken when leaving last page of a multipage screen, if the list screen you return to does not have the same forms connected in its own relationship. This way the relationship in effect when the data was entered, is used to process links, etc.