Making text for display spanning two cols work properly in conditions
When there may be only a single column displayed in the form. Tricky! Key is that forms are rendered in reverse sequence, subs first, then main form, so mainform settings would override. Need to check for the element in the screen being rendered, before we do the "reset" action to clean out the values stored in the session.