Removal of language setting in functions.php
Shouldn't be necessary! Some bilingual pages might get English? Or some constants might appear, in rare circumstances? Don't think this inclusion is actually necessary in this location any longer, and it should not be necessary even if there's a problem. Language files should be included as part of the controller type parts of the code execution, not the functions library.