Improved handling of page conditions for new entries
Now new entries are evaluated against {BLANK} as true. Previously, conditions were ignored for new entries. This may require existing screens that have conditions to add appropriate references to {BLANK} ie: where a page used to show up on new entires, but a condition applied to existing ones, such as, field = 'X', now that condition should go to the One Or More section, and a field = {BLANK} condition should be added, so that the page still meets the condition for new entries.