Adjust date searches to be timezone aware
If a user searches for {TODAY} and on the server {TODAY} is tomorrow or something like that, then we adjust the time accordingly so the semantics of the search match what the user typed. Except beware that data entered is always recorded with the now() timestamp of the server, so today starts and ends at midnight in the server's timezone. Ergo, server should really always be set to the timezone of the largest cohort of users. Or we should alter how we record the creation and mod times, so we get the local time of the user?? Time conversion just plain sucks all round.