Preserve specific value in linked lists with duplicates
Linked elements might contain multiple values with same human readable value. In these cases, isolate the specific copy that was selected, so we render with the exact value in the list that the user actually chose. Shouldn't actually matter, but can become visible as a problem when the list is disabled, because no value will show up because the underlying ID selected won't match anything in the list. So this fixes that.