Fixes to the order of subform defaults being written
Really, there's a problem with how defaults are gathered in general, because if defaults are dependent on prior determination of other defaults, then we're kind of screwed. With this and the prior commit, date defaults are not written if they are not determined, so when we do the second pass, they get determined based on defaults set in the first pass. Ugly!