Fixing tidy element lists so the save button works in save view popup
We used to not have them toggle the display CSS setting, but this meant the click interpretation in the DOM would detect checkboxes instead of save button. Now it shows/hides properly and button works, though the width of the section is not preserved, which is probably why we used opacity only, so that visually the change columns would be standard regardless of what's displayed... we could set a fixed width on that box though. And have.