PHP 8 fixes from Parlin
* Fix for undefined constant label for Procedure tab

This is for "Remove" button. It throws Fatal Error in PHP 8.1 causing the Form Setting page to not finish loading properly.

* Fix typo

* Check variable type before proceeding

These cause Fatal Error in PHP 8.1

* Tiny logic tidy up

Since the original error was because the value was an array, "not is array" was the default assumption. But really, this should only be done if the value is a string, because we're evaluating if the string contains "newvalue"

Co-authored-by: Julian Egelstaff <jegelstaff@users.noreply.github.com>