Avoiding infinite loops of derived value calls
If an on before save or on after save procedure called formulize_updateDerivedValues on the same form, then the writing of the derived values to the database, would trigger the updating of the derived values again, and again and again... Not anymore. We only process a derived value update if we're not in the middle of processing that derived value update already.