Improvements to custom button handling
Allows buttons to operate sometimes when there is no entry applicable until they are clicked, whereas previously they would not be available because we couldn't determine an entry at render time.

Also, buttons can be replaced with text, if the condition they apply to an entry is already met. This is experimental and needs to be fleshed out/completed with proper admin side controls for the text.