Update on change for element handles in derived values and text for display
Needs further extension to alter how refs to elements are stored in everywhere that the conditions UI is used, plus also in the advanceview list of elements in LOE settings, so that we rely on IDs instead. This needs careful work through the DB updates process, or backwards compatibility, or both, so we don't break existing systems when refs are changed.