Allowing dropdown filters on any field
Query for the values the user has access to, use those as the filter options. Includes support for <span class='formulize-filter-value'>Text</span> as a special code to declare a filter value if you're formatting lots of HTML in a derived value, for example, so that the filter choice can be simplified compared to the text that is displayed for the entire field on screen.