Refactoring Entry Locking Feature
Works properly on reloads, race conditions, modals... should be pretty bulletproof. But only kicks in properly on formulize list and form screens, including modals. In custom interfaces or future screen types the apparatus would have to be embedded in the new code. Fairly simple though, requires the on unload event in JS, and requires the delete locks jQuery to be sent on any in page ajax event that saves data/redraws the form. Also requires the lock token to be sent with any submission so the recreation of the elements will be able to remove the locks (to avoid race conditions with the on unload beacon)