Bug fix related to datepicker defaults when drawing subforms
Kind of silly because for legacy reasons we're going through the whole prep of rendering an entry that we're not going to render, before we get to actually rendering the subform, and in this case the datepicker defaults will already have been defined, but we need to throw them away so they can be defined again on the real page load! Ugh.