Better determination of whether values have changed before writing
Properly catches {WRITEASNULL} values and also properly handles NULL values in the element_values array, because isset returns false if a variable is set to NULL, argh! So loop through element values instead of existing values and we don't need to check if the element values array has the value in question. OK to iterate based on element values because all we were doing is removing things from element values anyway, so simpler to loop through that to begin with.