Make all code editors use highlighting
PHP tags need to be added to the start of all code blocks. And we need to remove the starting tags when we evaluate them.

A function handles removing the tags. A DB patch operation handles adding them. Not having tags is totally OK, but the highlighting won't kick in if you don't put a tag at the top of your code.

The editors that weren't code editors now are code editors thanks to class being added.

forms.php class file has been reorganized a little to be internally consistent in the order of functions.