Live derived value updates (#641)
If a derived value has a display conditional based on another element(s), and that element(s) are referenced in the derived value formula, then when the derived value is conditionally displayed/redisplayed, it will show a value based on the live value presently in the form on screen.

Such values are not saved to the database, they are only displayed. The computation is repeated when the entry is saved, and the values at that time are used to compute the derived value.

If a derived value formula is doing odd things and interacting with other data, those operations will proceed when the conditional display happens, same as if there is a save operation. This could be problematic, and people should only use conditional derived values when appropriate, and perform other operations affecting other data in on_before_save and on_after_save events, for example.

This feature could be made easier to use if we add some awareness/metadata to derived values to declare which elements they are referencing in code, and then we could automatically make them conditional based on those elements. Right now, the webmaster has to specifically add a display condition to the derived value for the live updates to take place. Typical display conditions would be "element-we-depend-on IS NOT {BLANK}"