Handle lookups filters on multi value elements
An improvement to commit 02087789d89a97a162989db40753c278ae61c79c. When an element has multiple values (stored with *=+*: by convention), then it must be searched in an odd way using LIKE or NOT LIKE. We interpret = and != as requiring this kind of operation since it doesn't make sense to do an equals search for the entire set of values. This in theory permits =X AND =Y to work for looking up checkboxes that have X and Y both checked.