Docs for clean functions for gathering data (#664)
The old functions have needed semantic cleanup for a while. Now...

getData -> gatherDataset
display -> getValue
NEW: getValue(raw: true) -- a new option! enabled by some of the 7.4 optimizations to how data is gathered vs prepped

And documentation for the three functions. Plus docs for buildScope.

Importantly, this also surfaces the ability to select only certain fields, which has been in getData for ages. But is not used generally because when building the entire interface, we don't know what fields someone might be referring to in code, so we get them all, just in case. We could be smarter about that with some static analysis of template code and derived values and stuff, but that's a lot of work! Regardless, end users of the functions should always select only the fields they want, to cut down on processing time, etc.

Commits:

* First pass at roughing in the functions, creating one doc file

* Work in progress on doc blocks and code cleanup

* Making changes through internal code to use new names

* Smarten up function name

* Adding docblock and clean up for buildScope function

* All done draft one

* Wording touchups

* Final touchups to the docs and getEntryIds, etc

* Small optimization in getting entry ids