Stop storing CSRF tokens in session (#722)
This should eliminate the last vestiges of "random" session loss, due to the session contents overflowing the storage size of the data field in the DB table. Problem made worse in utf8mb4, due to 4 bytes per char?

Token data is now encoded into filenames stored in a tokens folder off the root. Filenames have the format:

name_sessionId_tokenId_timeout

The name is the namespacing of the token (default is XOOPS_TOKEN). SessionId is the PHP session Id of the user. TokenId is the id value of the token, which when hashed appropriately will match a corresponding value passed from the client. Timeout is the time after which we should expire the token.

All operations of the libraries/icms/core/Security.php file are intact, the method of storage is simply files now, and not the $_SESSION. The files are zero bytes, just having names (long names, but no contents).